Add functions to initialize the LCD window
Add functions to initialize the LCD window, to read the binary file and load
character patterns, and to load the LCD's image.

Add functionality to close the window when the user clicks on the close button
on the window.

Since the library uses SDL2, for the LCD's screen we used the object
SDL_Renderer and for the image SDL_Texture instead of the surfaces. The reason
is that the SDL2 library recommends to use this objects.