Change pixel creation for the LCD
Previously the pixels for the LCD were created by getting surfaces. Then the
SDL_BlitSurface() was used to copy those surface into the screen.

However, according to the migration guide of the SDL2 library
(https://wiki.libsdl.org/SDL2/MigrationGuide#if_your_game_wants_to_blit_surfaces_to_the_screen)
the new way to create the pixels is to create textures from those surfaces and
copy the textures into the screen. This drawing method is recommended since the
pixels are not going to change, but also we are taking advantage of functions
that work directly with the GPU, which is faster.