Add function to write characters on the LCD
Add functions to write characters on the LCD, and to constantly refresh the
cursor, either because it's blinking or because it need to be moved to other
position.

From the structure GraphicUnit, the pointer *temp_screen was deleted since it's
not used in the code and it's purpose doesn't fit with the current drawing
method use to draw the LCD pixels on the screen.