Update logic to create the SDL window and renderer
Previously, the SDL window and renderer was done inside the main application.
However, this is not a good practice and instead the logic was moved to the
library.

Now the library creates the window and the renderer, and also initializes the
LCDSim structure. Doing it this way hides the internal logic from the user.