Move HD44780 functions into a new source file
Move the hardware emulation functions into a new source file which is going to
be inside a library's sub-directory. The reason for this changes is to have a
modular code, instead of having one single file with all the different features
in it.

Also, since there are macros used among library's source files, a new header
file was created with all the macros and definitions of variables that are
share inside the library's files.