Change logic to draw the pixel on the LCD
The previous way to draw the pixels was inefficient and required a lot of
variables.

With the new implementation only the buffer with the pixels is needed to display
the pixels. The module hd44780.c updates this buffer and sends it back to the
module lcdsim.c. Then, this module reads the buffer and selects between the two
colors, green or black, to draw on the LCD.

The enumaration for the LCD colors was changed so it is coherent with the actual
behavior of the chip. This means that Green (0) represent a pixel OFF, and
Black (1) represents a pixel ON.

Other major udpate is that all the logic related to SDL2 is inside the lcdsim.c
module. The reason is to separate both functionalities; lcdsim.c is in charge of
drawing the simulated LCD, and hd44780.c is in charge of simulating the logic of
the IC.