Remove functions related to the API for LCD
The API functions to control the LCD were removed since a third party library is
going to be used. This new library will be modified so it calls the LCDSim
functions.

The third party library will be used as an example on how to import a library to
this project and be used to simulate the micrcontroller using teh functions of
the imported library.