Move the instructions to copy the resource files
The instructions to copy the files inside the directory res/ were moved to the
top CMakeLists.txt.

The reason is that it makes more sense that the main cmake file deals with the
resources for the project rather than the cmake file of the exmaple application
doing the copy of the resources.