Set library version
Previously the cmake_project_version was the same for the library, although this
is going to be true for both, it was decided to separate the version of the
library from the project. The reason is because the library is going to be used
in other projects so it should keep the current version of the library.

With the previous approach, the version was shared to library but in the situation
where the librarie's directory is copied to other project, it will inherit the
version of that new project, which is not desired.