Define CMAKE macros for files in res/ dir
In commit 000c55 was defined a solution to fix the problem of the program not
finding the resources due to a wrong path. However, that didn't work and after
a clean up of the project and rebuild it failed again.

In this commit finally is solved this issue. Instead of hardcoding the path to
the resources, a CMAKE macros are used to refer to that path of those resources.
So now the program knows where to find the .bin and .bmp files because we are
passing it, through the macros, the absolut path of them. But more important,
since this a static library, it doesn't matter from where we are calling the
functions to load these files the program will always know where to find them.